+++ pi1/class.tx_realty_pi1_ImageThumbnailsView.php
@@ -1,5 +1,7 @@
 <?php
 
+use TYPO3\CMS\Core\Resource\FileReference;
+
 /**
  * This class renders the images for one realty object as thumbnails.
  *
@@ -26,7 +28,45 @@ class tx_realty_pi1_ImageThumbnailsView extends tx_realty_pi1_FrontEndView
     {
         $this->showUid = (int)$piVars['showUid'];
 
-        return ($this->renderLegacyImages() > 0) ? $this->getSubpart('FIELD_WRAPPER_IMAGETHUMBNAILS') : '';
+        $numberOfRenderedImages = $this->renderImages();
+        if ($numberOfRenderedImages === 0) {
+            $numberOfRenderedImages = $this->renderLegacyImages();
+        }
+
+        return $numberOfRenderedImages > 0 ? $this->getSubpart('FIELD_WRAPPER_IMAGETHUMBNAILS') : '';
+    }
+
+    /**
+     * Creates all images that are attached to the current record and puts them in their particular subparts.
+     *
+     * @return int the total number of rendered images
+     *
+     * @throws \Tx_Oelib_Exception_NotFound
+     */
+    private function renderImages()
+    {
+        $enableLightbox = $this->getPluginConfiguration()->getAsBoolean('enableLightbox');
+        /** @var \tx_realty_Mapper_RealtyObject $realtyObjectMapper */
+        $realtyObjectMapper = \Tx_Oelib_MapperRegistry::get(\tx_realty_Mapper_RealtyObject::class);
+        /** @var \tx_realty_Model_RealtyObject $realtyObject */
+        $realtyObject = $realtyObjectMapper->find($this->getUid());
+        $images = $realtyObject->getJpegAttachments();
+        if (\count($images) === 0) {
+            return 0;
+        }
+
+        $result = '';
+        foreach ($images as $image) {
+            $currentImage = $enableLightbox
+                ? $this->createLightboxThumbnail($image)
+                : $this->createThumbnail($image);
+            $this->setMarker('one_image_tag', $currentImage);
+            $result .= $this->getSubpart('ONE_IMAGE_CONTAINER');
+        }
+
+        $this->setSubpart('ONE_IMAGE_CONTAINER', $result);
+
+        return count($images);
     }
 
     /**
@@ -49,7 +89,8 @@ class tx_realty_pi1_ImageThumbnailsView extends tx_realty_pi1_FrontEndView
         /** @var \tx_realty_Model_Image $image */
         foreach ($realtyObject->getImages() as $image) {
             $currentImage = $enableLightbox
-                ? $this->createLegacyLightboxThumbnail($image) : $this->createLegacyThumbnail($image);
+                ? $this->createLegacyLightboxThumbnail($image)
+                : $this->createLegacyThumbnail($image);
             $this->setMarker('one_image_tag', $currentImage);
             $result .= $this->getSubpart('ONE_IMAGE_CONTAINER');
         }
@@ -86,7 +127,6 @@ class tx_realty_pi1_ImageThumbnailsView extends tx_realty_pi1_FrontEndView
      * Creates a Lightbox thumbnail of $image sized as per the configuration.
      *
      * @param \tx_realty_Model_Image $image
-     *        the image to render
      *
      * @return string
      *         image tag wrapped in a Lightbox link, will not be empty
@@ -99,6 +139,62 @@ class tx_realty_pi1_ImageThumbnailsView extends tx_realty_pi1_FrontEndView
             'altText' => $image->getTitle(),
             'titleText' => $image->getTitle(),
             'file' => \tx_realty_Model_Image::UPLOAD_FOLDER . $image->getFileName(),
+            'file.' => [
+                'maxW' => $configuration->getAsInteger('lightboxImageWidthMax'),
+                'maxH' => $configuration->getAsInteger('lightboxImageHeightMax'),
+            ],
+        ];
+        $imageWithTag = $this->cObj->cObjGetSingle('IMAGE', $imageConfiguration);
+
+        $imagePath = [];
+        \preg_match('/src="([^"]*)"/', $imageWithTag, $imagePath);
+        $fullSizeImageUrl = $imagePath[1];
+
+        $linkAttribute = ' data-lightbox="objectGallery" data-title="' .
+            \htmlspecialchars($image->getTitle(), ENT_COMPAT | ENT_HTML5) . '"';
+
+        return '<a href="' . $fullSizeImageUrl . '"' . $linkAttribute . '>' . $thumbnailTag . '</a>';
+    }
+
+    /**
+     * Creates a thumbnail (without Lightbox) of $image sized as per the configuration.
+     *
+     * @param FileReference $image
+     *
+     * @return string image tag, will not be empty
+     */
+    protected function createThumbnail(FileReference $image)
+    {
+        $configuration = $this->getPluginConfiguration();
+        $imageConfiguration = [
+            'altText' => $image->getTitle(),
+            'titleText' => $image->getTitle(),
+            'file' => $image->getPublicUrl(),
+            'file.' => [
+                'width' => $configuration->getAsInteger('singleImageMaxX') . 'c',
+                'height' => $configuration->getAsInteger('singleImageMaxY') . 'c',
+            ],
+        ];
+
+        return $this->cObj->cObjGetSingle('IMAGE', $imageConfiguration);
+    }
+
+    /**
+     * Creates a Lightbox thumbnail of $image sized as per the configuration.
+     *
+     * @param FileReference $image
+     *
+     * @return string
+     *         image tag wrapped in a Lightbox link, will not be empty
+     */
+    protected function createLightboxThumbnail(FileReference $image)
+    {
+        $thumbnailTag = $this->createThumbnail($image);
+        $configuration = $this->getPluginConfiguration();
+        $imageConfiguration = [
+            'altText' => $image->getTitle(),
+            'titleText' => $image->getTitle(),
+            'file' => $image->getPublicUrl(),
             'file.' => [
                 'maxW' => $configuration->getAsInteger('lightboxImageWidthMax'),
                 'maxH' => $configuration->getAsInteger('lightboxImageHeightMax'),